minor change Rental (Long / integer)

+++ src/main/java/com/chatop/api/service/RentalService.java
@@ -1,15 +1,11 @@
-/* V1 construit avec le cours */
-
 package com.chatop.api.service;
 
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.chatop.api.model.Rental;
 import com.chatop.api.repository.RentalRepository;
 
+import org.springframework.stereotype.Service;
 import lombok.Data;
 
 @Data
@@ -19,7 +15,7 @@ public class RentalService {
     @Autowired
     private RentalRepository rentalRepository;
 
-    public Optional<Rental> getRental(final Long id) {
+    public Optional<Rental> getRental(final Integer id) {
         return rentalRepository.findById(id);
     }
 
@@ -27,13 +23,14 @@ public class RentalService {
         return rentalRepository.findAll();
     }
 
-    public void deleteRental(final Long id) {
+    public void deleteRental(final Integer id) {
         rentalRepository.deleteById(id);
     }
 
     public Rental saveRental(Rental rental) {
         Rental savedRental = rentalRepository.save(rental);
         return savedRental;
-    }
+    }  
 
 }
+
